+++ app/Main.hs
@@ -1,4 +1,4 @@
-{- | word-count: count lines/words/characters in files.
+{- | word-count: count lines/words/characters/bytes in files.
      Main.hs
 -}
 module Main (main) where
@@ -14,6 +14,7 @@ data Flags = Flags
     , countLines :: Bool        -- Boolean flag to count lines in files
     , countWords :: Bool        -- Boolean flag to count words in files
     , countChars :: Bool        -- Boolean flag to count characters in files
+    , countBytes :: Bool        -- Boolean flag to count bytes in files
     }
 
 {- | A parser built applicative style for optparse-applicative. -}
@@ -33,6 +34,10 @@ flags = Flags
       (long "countChars"
       <> short 'c'
       <> help "Count characters in a file")
+    <*> switch
+      (long "countBytes"
+      <> short 'b'
+      <> help "Count file size in bytes")
 
 {- | Main function. -}
 main :: IO ()
@@ -40,95 +45,167 @@ main = countFileStats =<< execParser opts
   where
     opts = info (flags <**> helper)
       (fullDesc
-       <> progDesc "Count words in files."
+       <> progDesc "Count lines/words/characters/bytes in file(s)."
        <> header "word-count")
 
-{- | Recursively count a list of file paths for occurances of lines, words,
-     and characters.
--}
+{- | Print the line cout, word count, character count and size of file(s). -}
 countFileStats :: Flags -> IO ()
 
 {- | Edge case: the empty FilePath list, return () -}
-countFileStats (Flags [] _ _ _) = pure ()
+countFileStats (Flags [] _ _ _ _) = pure ()
 
-{- | File(s) were given but no flags, run all counting functions. -}
-countFileStats (Flags fP False False False) = do
+{- | File(s) were given but no flags, print all counts in file(s). -}
+countFileStats (Flags fP False False False False) = do
     let file = head fP
     putHeader file
-    withFile file ReadMode $ \handle -> do
-        contents <- readFileContents handle
-        putFileCounts contents
-        countFileStats (Flags (drop 1 fP) False False False)
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putFileCounts fileStats
+        countFileStats (Flags (drop 1 fP) False False False False)
 
-{- | countLines flag was given so run the count lines function. -}
-countFileStats (Flags fP cL False False) = do
+{- | countLines flag was given so print the line count in file(s). -}
+countFileStats (Flags fP cL False False False) = do
     let file = head fP
     putHeader file
-    withFile file ReadMode $ \handle -> do
-        contents <- readFileContents handle
-        putLineCount contents
-        countFileStats (Flags (drop 1 fP) cL False False)
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putLineCount (lc' fileStats)
+        countFileStats (Flags (drop 1 fP) cL False False False)
 
-{- | countWords flag was given so run the count words function. -}
-countFileStats (Flags fP False cW False) = do
+{- | countWords flag was given so print the word count in file(s). -}
+countFileStats (Flags fP False cW False False) = do
     let file = head fP
     putHeader file
-    withFile file ReadMode $ \handle -> do
-        contents <- readFileContents handle
-        putWordCount contents
-        countFileStats (Flags (drop 1 fP) False cW False)
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putWordCount (wc' fileStats)
+        countFileStats (Flags (drop 1 fP) False cW False False)
 
-{- | countChars flag was given so run the count characters function. -}
-countFileStats (Flags fP False False cC) = do
+{- | countChars flag was given so print the character count in file(s). -}
+countFileStats (Flags fP False False cC False) = do
     let file = head fP
     putHeader file
-    withFile file ReadMode $ \handle -> do
-        contents <- readFileContents handle
-        putCharCount contents
-        countFileStats (Flags (drop 1 fP) False False cC)
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putCharCount (wc' fileStats)
+        countFileStats (Flags (drop 1 fP) False False cC False)
 
-{- | Count lines and words in files. -}
-countFileStats (Flags fP cL cW False) = do
+{- | countBytes flag was given so print the byte count in file(s). -}
+countFileStats (Flags fP False False False cB) = do
     let file = head fP
     putHeader file
-    withFile file ReadMode $ \handle -> do
-        contents <- readFileContents handle
-        putLineCount contents
-        putWordCount contents
-        countFileStats (Flags (drop 1 fP) cL cW False)
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putByteCount (bc' fileStats)
+        countFileStats(Flags (drop 1 fP) False False False cB)
 
-{- | Count lines and characters in files. -}
-countFileStats (Flags fP cL False cC) = do
+{- | Print the line and word counts in file(s). -}
+countFileStats (Flags fP cL cW False False) = do
     let file = head fP
     putHeader file
-    withFile file ReadMode $ \handle -> do
-        contents <- readFileContents handle
-        putLineCount contents
-        putCharCount contents
-        countFileStats (Flags (drop 1 fP) cL False cC)
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putLineCount (lc' fileStats)
+        putWordCount (wc' fileStats)
+        countFileStats (Flags (drop 1 fP) cL cW False False)
 
-{- | Count words and characters in files. -}
-countFileStats (Flags fP False cW cC) = do
+{- | Print the line and character count in file(s). -}
+countFileStats (Flags fP cL False cC False) = do
     let file = head fP
     putHeader file
-    withFile file ReadMode $ \handle -> do
-        contents <- readFileContents handle
-        putWordCount contents
-        putCharCount contents
-        countFileStats (Flags (drop 1 fP) False cW cC)
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putLineCount (lc' fileStats)
+        putCharCount (cc' fileStats)
+        countFileStats (Flags (drop 1 fP) cL False cC False)
 
-{- | All flags were given, run all counting functions. -}
-countFileStats (Flags fP cL cW cC) = do
+{- | Print the word and character count in file(s). -}
+countFileStats (Flags fP False cW cC False) = do
     let file = head fP
     putHeader file
-    withFile file ReadMode $ \handle -> do
-        contents <- readFileContents handle
-        putFileCounts contents
-        countFileStats (Flags (drop 1 fP) cL cW cC)
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putWordCount (wc' fileStats)
+        putCharCount (cc' fileStats)
+        countFileStats (Flags (drop 1 fP) False cW cC False)
+
+{- | Print the line count and byte size of file(s). -}
+countFileStats (Flags fP cL False False cB) = do
+    let file = head fP
+    putHeader file
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putLineCount (lc' fileStats)
+        putByteCount (bc' fileStats)
+        countFileStats (Flags (drop 1 fP) cL False False cB)
+
+{- | Print the word count and byte size of file(s). -}
+countFileStats (Flags fP False cW False cB) = do
+    let file = head fP
+    putHeader file
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putWordCount (wc' fileStats)
+        putByteCount (bc' fileStats)
+        countFileStats (Flags (drop 1 fP) False cW False cB)
+
+{- | Print the character count and byte size of file(s). -}
+countFileStats (Flags fP False False cC cB) = do
+    let file = head fP
+    putHeader file
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putCharCount (cc' fileStats)
+        putByteCount (bc' fileStats)
+        countFileStats (Flags (drop 1 fP) False False cC cB)
+
+{- | Print the line, word, character count and byte size of file(s). -}
+countFileStats (Flags fP cL cW False cB) = do
+    let file = head fP
+    putHeader file
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putLineCount (lc' fileStats)
+        putWordCount (wc' fileStats)
+        putByteCount (bc' fileStats)
+        countFileStats (Flags (drop 1 fP) cL cW False cB)
+
+{- | Print the line, character count and byte size of file(s). -}
+countFileStats (Flags fP cL False cC cB)= do
+    let file = head fP
+    putHeader file
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putLineCount (lc' fileStats)
+        putCharCount (cc' fileStats)
+        putByteCount (bc' fileStats)
+        countFileStats (Flags (drop 1 fP) cL False cC cB)
+
+{- | Print the word, character count and byte size of file(s). -}
+countFileStats (Flags fP False cW cC cB) = do
+    let file = head fP
+    putHeader file
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putWordCount (wc' fileStats)
+        putCharCount (cc' fileStats)
+        putByteCount (bc' fileStats)
+        countFileStats (Flags (drop 1 fP) False cW cC cB)
+
+{- | All flags were given, print all counts in file(s). -}
+countFileStats (Flags fP cL cW cC cB) = do
+    let file = head fP
+    putHeader file
+    withBinaryFile file ReadMode $ \handle -> do
+        fileStats <- readFileStats handle
+        putFileCounts fileStats
+        countFileStats (Flags (drop 1 fP) cL cW cC cB)
 
 {- | Print a header for the file being processed. -}
 putHeader :: FilePath -> IO ()
 putHeader file = do
+    replicateM_ (length file) (putStr "=")
+    putStr "\n"
     putStrLn file
     replicateM_ (length file) (putStr "=")
     putStr "\n"